separating functionalities to make it more clear
